+++ src/main/java/shop/gitit/member/service/LoginService.java
@@ -13,6 +13,7 @@ import shop.gitit.member.repository.RedisRepository;
 import shop.gitit.member.service.dto.response.GithubUserInfo;
 import shop.gitit.member.service.dto.response.LoginResDto;
 import shop.gitit.member.service.dto.response.OAuthTokenResponse;
+import shop.gitit.member.service.port.in.JoinUsecase;
 import shop.gitit.member.service.port.in.LoginUsecase;
 import shop.gitit.member.service.port.out.OAuthWebClient;
 
@@ -22,7 +23,7 @@ import shop.gitit.member.service.port.out.OAuthWebClient;
 @Transactional
 public class LoginService implements LoginUsecase {
 
-    private final JoinService joinService;
+    private final JoinUsecase joinUsecase;
     private final JwtTokenProvider jwtTokenProvider;
     private final OAuthWebClient oAuthWebClient;
     private final MemberRepository memberRepository;
@@ -32,14 +33,10 @@ public class LoginService implements LoginUsecase {
     public LoginResDto login(String code) {
         OAuthTokenResponse tokenResponse = oAuthWebClient.getToken(code);
         GithubUserInfo githubUserInfo = oAuthWebClient.getUserProfile(tokenResponse);
-        Member member = memberRepository.findByGithubId(githubUserInfo.getGithubId()).orElse(null);
-        member = joinIfNewMember(githubUserInfo, member);
-        updateProfileImg(githubUserInfo, member);
+        Member member = joinUsecase.join(githubUserInfo);
         JwtToken jwtToken = jwtTokenProvider.createToken(member.getId());
         redisRepository.save(
-                RefreshToken
-                        .builder() // TODO redis에 저장까지 완료한 후, 이후 로직에서 에러가 발생하여 트랜잭션 롤백이 발생한다면 RTK는
-                        // 삭제되어야 한다.
+                RefreshToken.builder()
                         .memberId(member.getId())
                         .token(jwtToken.getRefreshToken())
                         .build());
@@ -48,18 +45,5 @@ public class LoginService implements LoginUsecase {
                 .refreshToken(jwtToken.getRefreshToken())
                 .accessToken(jwtToken.getAccessToken())
                 .build();
-    }
-
-    private Member joinIfNewMember(GithubUserInfo githubUserInfo, Member member) {
-        if (member == null) {
-            member = joinService.join(githubUserInfo);
-        }
-        return member;
-    }
-
-    private void updateProfileImg(GithubUserInfo githubUserInfo, Member member) {
-        if (!member.getProfile().getProfileImg().equals(githubUserInfo.getProfileImg())) {
-            member.getProfile().updateProfileImg(githubUserInfo.getProfileImg());
-        }
     }
 }
