+++ src/main/java/shop/gitit/member/service/port/in/JoinUsecase.java
@@ -1,3 +1,9 @@
 package shop.gitit.member.service.port.in;
 
-public interface JoinUsecase {}
+import shop.gitit.member.domain.Member;
+import shop.gitit.member.service.dto.response.GithubUserInfo;
+
+public interface JoinUsecase {
+
+    Member join(GithubUserInfo githubUserInfo);
+}
